feat : 신발추가 / 목록조회 API 연결

+++ RunTaeho/UIViews/MyInfo/Shoes/Dtos/AddShoeViewDto.swift
@@ -10,5 +10,5 @@ import Foundation
 struct AddShoeViewDto {
     let brand: String
     let model: String
-    let targetDistance: Double
+    let targetDistance: Int
 }
